#291 - create merger module add prefix formatting function
