+++ web/controllers/helpers.ex
@@ -2,16 +2,6 @@ defmodule Karma.Controllers.Helpers do
 
   alias Karma.RedisCli
 
-  alias Karma.Repo
-
-  def get_data_for_merge(offer) do
-    %{user: Map.from_struct(Repo.get(Karma.User, offer.user_id)),
-      project: Map.from_struct(Repo.get(Karma.Project, offer.project_id)),
-      offer: Map.from_struct(Repo.get(Karma.Offer, offer.id)),
-      startpack: Map.from_struct(Repo.get_by(Karma.Startpack, user_id: offer.user_id))
-    }
-  end
-
   def user_startpack(user) do
     Ecto.assoc(user, :startpacks)
   end
